+++ tests/test_models/test_sarimax_model.py
@@ -38,6 +38,23 @@ def test_sarimax_forecaster_run_with_reg(example_reg_tsds):
     assert len(res) == 14
 
 
+def test_sarimax_forececaster_run_with_reg_custom_order(example_reg_tsds):
+    """
+    Given: I have dataframe with 2 segments
+    When: Sarimax have non standard `order` param
+    Then: I get 7 periods per dataset as a forecast
+    """
+    horizon = 7
+    model = SARIMAXModel(order=(3, 1, 0))
+    model.fit(example_reg_tsds)
+    future_ts = example_reg_tsds.make_future(future_steps=horizon)
+    res = model.forecast(future_ts)
+    res = res.to_pandas(flatten=True)
+
+    assert not res.isnull().values.any()
+    assert len(res) == 14
+
+
 def test_prediction_interval_run_insample(example_tsds):
     model = SARIMAXModel()
     model.fit(example_tsds)